change naming according to the ubiquitous language dictionary
