--- fleetManagement/build.gradle.kts
@@ -1,3 +1,0 @@
-subprojects {
-    group = "com.polyakovworkbox.stringconcatcourse.fleetmanagement"
-}
