--- fleetManagement/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftAddedToTheFleetDomainEvent.kt
@@ -1,5 +1,0 @@
-package com.polyakovworkbox.stringconcatcourse.fleetManagement.domain.aircraft
-
-import com.polyakovworkbox.stringconcatcourse.common.types.base.DomainEvent
-
-data class AircraftAddedToTheFleetDomainEvent(val aircraftId: AircraftId) : DomainEvent()
