--- fleetManagement/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftAlreadyExistsChecker.kt
@@ -1,5 +1,0 @@
-package com.polyakovworkbox.stringconcatcourse.fleetManagement.domain.aircraft
-
-interface AircraftAlreadyExistsChecker {
-    fun check(registrationNumber: AircraftRegistrationNumber): Boolean
-}
