--- fleetManagement/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftId.kt
@@ -1,7 +1,0 @@
-package com.polyakovworkbox.stringconcatcourse.fleetManagement.domain.aircraft
-
-data class AircraftId(val value: Long)
-
-interface AircraftIdGenerator {
-    fun generate(): AircraftId
-}
