+++ leasing/build.gradle.kts
@@ -1,0 +1,3 @@
+subprojects {
+    group = "com.polyakovworkbox.stringconcatcourse.leasing"
+}
