+++ leasing/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftAcquiredDomainEvent.kt
@@ -1,0 +1,5 @@
+package com.polyakovworkbox.stringconcatcourse.leasing.domain.aircraft
+
+import com.polyakovworkbox.stringconcatcourse.common.types.base.DomainEvent
+
+data class AircraftAcquiredDomainEvent(val aircraftId: AircraftId) : DomainEvent()
