+++ leasing/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftAlreadyExistsChecker.kt
@@ -1,0 +1,5 @@
+package com.polyakovworkbox.stringconcatcourse.leasing.domain.aircraft
+
+interface AircraftAlreadyExistsChecker {
+    fun check(registrationNumber: AircraftRegistrationNumber): Boolean
+}
