+++ leasing/domain/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/fleetManagement/domain/aircraft/AircraftId.kt
@@ -1,0 +1,7 @@
+package com.polyakovworkbox.stringconcatcourse.leasing.domain.aircraft
+
+data class AircraftId(val value: Long)
+
+interface AircraftIdGenerator {
+    fun generate(): AircraftId
+}
