--- mainApp/src/main/kotlin/com/polyakovworkbox/stringconcatcourse/configuration/FleetManagementConfiguration.kt
@@ -1,12 +1,0 @@
-package com.polyakovworkbox.stringconcatcourse.configuration
-
-/*import com.polyakovworkbox.stringconcatcourse.fleetManagement.domain.aircraft.Aircraft
-import org.springframework.context.annotation.Bean*/
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@Suppress("TooManyFunctions")
-class FleetManagementConfiguration /*{
-    @Bean
-    fun getAircraft() = Aircraft()
-}*/
