+++ settings.gradle.kts
@@ -1,8 +1,9 @@
 rootProject.name = "airlinecompanyapplication"
 
-include(":fleetManagement:domain")
+include(":leasing:domain")
 
 include(":mainApp")
 include(":common:types")
 include(":integration")
+
