+++ src/test/kotlin/com/stringconcat/ddd/fitnes/MetricsGuard.kt
@@ -7,29 +7,27 @@ import com.tngtech.archunit.junit.AnalyzeClasses
 import com.tngtech.archunit.junit.ArchTest
 import com.tngtech.archunit.library.metrics.ArchitectureMetrics
 import com.tngtech.archunit.library.metrics.MetricsComponents
-import io.kotest.core.annotation.Ignored
 
-@Ignored
 @AnalyzeClasses(packages = ["com.polyakovworkbox.stringconcatcourse"])
 class MetricsGuard {
 
     @ArchTest
-    fun `uncle bob metrics for fleetManagement`(importedClasses: JavaClasses) {
+    fun `uncle bob metrics for leasing`(importedClasses: JavaClasses) {
         val packages: Set<JavaPackage> = importedClasses.getPackage(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement").getSubpackages()
+            "com.polyakovworkbox.stringconcatcourse.leasing").getSubpackages()
         val components: MetricsComponents<JavaClass> = MetricsComponents.fromPackages(packages)
 
         val metrics = ArchitectureMetrics.componentDependencyMetrics(components)
 
         println("Ce: " + metrics.getEfferentCoupling(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement.domain"))
+            "com.polyakovworkbox.stringconcatcourse.leasing.domain"))
         println("Ca: " + metrics.getAfferentCoupling(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement.domain"))
+            "com.polyakovworkbox.stringconcatcourse.leasing.domain"))
         println("I: " + metrics.getInstability(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement.domain"))
+            "com.polyakovworkbox.stringconcatcourse.leasing.domain"))
         println("A: " + metrics.getAbstractness(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement.domain"))
+            "com.polyakovworkbox.stringconcatcourse.leasing.domain"))
         println("D: " + metrics.getNormalizedDistanceFromMainSequence(
-            "com.polyakovworkbox.stringconcatcourse.fleetManagement.domain"))
+            "com.polyakovworkbox.stringconcatcourse.leasing.domain"))
     }
 }
